Modularize system config
This commit fixes one thing from
https://github.com/skogsbrus/os/issues/14, namely modular system
components.

Everything is put under `sys/`, regardless if it's a service or app.
Might move things around later if it makes sense, but the current config
is small enough to keep it in one place.

This structure lets me configure everything declaratively from the top:

1. flake.nix decides which host to configure (based on host name)
2. hosts/<name>/host.nix loads hosts/<name>/hardware.nix and all system
   modules from `sys/default.nix`. Custom stuff is kept under my own
   namespace, skogsbrus.